Improve PHP 7 compatibility by handling highlight_string exceptions.
Code that's valid in PHP 5 but invalid in PHP 7 will cause ParseException
objects to be thrown by highlight_string(). We currently have one example of
this in php-doc-en, but will presumably have more as more PHP 7 documentation
is written.

This change will ensure that we handle this as gracefully as we can: instead of
erroring out, we'll at least include the code without syntax highlighting.