Add support for properly rendering nullable types
While it is already possible to write types with the nullable type mark
(`?`) (and some extensions already make use of this), these types are
not linked to the respective classes or interfaces.  So we make sure
that the type lookup ignores the question mark, and we put it outside
the link.

This is a bit hackish, because it would be better to introduce proper
support for union types; however, these are only supported as of
DocBook 5.2, and since we're still using 4.5, we would have to upgrade
first.