Support union types
For now we stick with the union type syntax introduced with PHP 8,
because proper union type markup[1] is only available as of DocBook
5.2, and we're still using version 4.5.  We do not support the nullable
type syntax (i.e. ? prefix) which has been introduced with PHP 7.1,
because this is just a shortcut for union types, and may not be
supported by future DocBook versions.

[1] <https://tdg.docbook.org/tdg/5.2/type.html>