Support variadic parameters
So far, PhD does not support variadic parameters.  Instead the manual
follows the convention to use `...` as parameter name, which is then
rendered as `$...` in the docs.  However, as of PHP 5.6.0 variadic
parameters are supported using the syntax `...$args`; therefore, we are
going to support this notation for PhD.

While DocBook supports a `<varargs>` element, this cannot be used in
combination with named parameters.  Therefore, we use the `role`
attribute of the `<parameter>` element to declare variadic parameters.
Since the `role` attribute is already used to designate reference
parameters, we support the following values now: `"reference"`,
`"variadic"` and `"reference variadic"`; additional whitespace as well
as swapping the order of the words is supported.