Add total number of functions and methods to the index
Apparently, users are interested in this number; there are constantly
new user notes which mention this number (a lot of these have been
deleted in the past).  This patch would display the total number of
functions and methods, so users won't need to calculate for themselves.

As it is now, i18n is completely missing.  We would need to define an
XML entity in language-snippets.ent for this.