Remove html_entity_decode when getting refentry names from the database (#131)
Refentry names are taken from xml IDs and stored in the database during indexing. These are then retrieved and used for linking to functions and methods. Do not use html_entity_decode on these names as these IDs should not have entities in them to begin with. This change aligns this function with the function/method linking functions that only do strtolower as well.

Co-authored-by: haszi <haszika80@gmail.com>